Tile & TileHandler methods
This commit disables the usage of the IsoGrid script and adds two new
ones:
Tile: A script that is added to the Tile Prefab, which is later meant to
have a cost value used for the pathfinding.
TileHandler: A script that initializes a TileMap[x,y] and instantiates
all the tiles and populates the TileMap with the new tiles.